rename: trainer info 경로 변경
