+++ src/components/Common/Card/TrainerInfo/Favorite.tsx
