+++ src/components/Common/Card/TrainerInfo/Rating.tsx
