+++ src/components/Common/Card/TrainerInfo/TrainerInfo.tsx
@@ -1,10 +1,10 @@
 import { ic_profile_default_md } from "@/imageExports";
 import clsx from "clsx";
 import Image from "next/image";
+import { VerticalLine } from "../Line";
 import Experience from "./Experience";
 import Favorite from "./Favorite";
 import LessonCount from "./LessonCount";
-import { VerticalLine } from "./Line";
 import Rating from "./Rating";
 
 const container = clsx(
